Все функции вынесены в библиотеку functions. В main осталась только логика, убраны лишние импорты.
