+++ funcions.py
@@ -1,4 +1,5 @@
 import time
+import random
 
 HARD_LEVEL_ATTEMPTS = 5
 EASY_LEVEL_ATTEMPTS = 1
@@ -59,7 +60,41 @@ def greeting(number_of_players):
         time.sleep(1.5)
         print('Вам предстоит его отгадать.')
         time.sleep(1.5)
-    
+
+def multiplayer():
+    players = funcions.add_player(count_players, 0 )
+    for name in players:
+        #Для каждого игрока будут разные случайные числа
+        print(f'Ход игрока {name}')
+        random_number = random.randint(1, 100)
+        user_guess = 0
+        while user_guess != random_number:
+            players[name] += 1
+            print(f'Количество совершенных ходов: {players[name]}.')
+            user_guess = int(input('Введите ваше число: '))
+            funcions.checking_guess_multiplayer(user_guess, random_number)
+  
+    #Превращаем словарь в список и сортируем по кол-ву попыток.
+    pairs = list(players.items())
+    pairs.sort(key=lambda pairs: pairs[1])
+    print(f'Побеждает {pairs[0][0]}. Количество ходов {pairs[0][1]}')
+
+
+
+
+def single_player():
+    random_number = random.randint(1, 100)
+    turns = funcions.choose_difficulty()
+    user_guess = 0
+    while  user_guess != random_number:
+        print(f'У вас осталось {turns} попыток')
+        user_guess = int(input('Введите ваше число: '))
+        funcions.checking_guess(user_guess, random_number) # проверяем ввод
+        turns -=1 #уменьшаем количество доступных попыток
+        while turns: # пока количетсво попыток больше или равно 1 - играем.
+            print('Не угадали, давайте еще раз.')
+        print(f'Вы так и не смогли отгадать число {random_number}. Вы проиграли.')
+
